KMake: Fix executing defconfig every time
Adding "kconfig" target dependency to "$(KCONFIG)" target caused the
"$(KCONFIG)" to be executed every time.  Rewrite the entire logic
to solve this - at least that's the only option I see, if the user kmake
is included at the beginning of the file which makes :: useless for
adding on top of defconfig.

Now the usage depends on user kmake variables CONFIG_DISABLE and
CONFIG_ENABLE.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>